requirements.txt
gradio==4.*
moviepy==1.0.3
imageio-ffmpeg
numpy<2.0
opencv-python-headless
pytubefix
pydub
pysrt
faster-whisper
google-generativeai
openai>=1.35.0
